Pre-allocate Memory in `codec.Packer` (#280)
* update codec to allow for initial memory allocation

* add Size() to chain dependency

* update chain/result

* add TotalSize util

* update transaction

* add size to actions

* add auth sizes

* update gossiper

* restore message buffer

* update rpc

* update warp manager

* re-generate mocks

* move BoolLen with rest of consts

* fix codec tests

* passing all unit tests

* make sure to add overhead for optional bits

* use CummSize

* fix integration tests

+++ gossiper/proposer.go
@@ -158,7 +158,7 @@ func (g *Proposer) ForceGossip(ctx context.Context) error {
 	// Gossip highest paying txs
 	var (
 		txs   = []*chain.Transaction{}
-		size  = uint64(0)
+		size  = 0
 		start = time.Now()
 		now   = start.UnixMilli()
 		r     = g.vm.Rules(now)
@@ -220,7 +220,7 @@ func (g *Proposer) ForceGossip(ctx context.Context) error {
 
 			// Gossip up to [consts.NetworkSizeLimit]
 			txSize := next.Size()
-			if txSize+size > uint64(g.cfg.GossipMaxSize) {
+			if txSize+size > g.cfg.GossipMaxSize {
 				return false, true, false, nil
 			}
 			txs = append(txs, next)
